+++ lib-model-consume/src/main/java/com/uber/nullaway/libmodel/StubxFileWriter.java
@@ -1,4 +1,4 @@
-package com.uber.nullaway;
+package com.uber.nullaway.libmodel;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
